fix: Move switch dialog to separate file for unified callback logic
