+++ packages/app_center/lib/snapd/snapd.dart
@@ -1,5 +1,6 @@
 export 'snap_action.dart';
 export 'snap_category_enum.dart';
+export 'snap_channel_switch_dialog.dart';
 export 'snap_data.dart';
 export 'snap_l10n.dart';
 export 'snap_launcher.dart';
